# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RFA knowledge server — /ask
  summary: desk(대응 에이전트) ↔ 지식 서버 계약
  description: 'head → broker → task agent → censor 를 수행하는 `ask()` 하나를 노출한다. `audience` 가 검열 프로파일을 정한다(public→public,
    company/self→internal). 동기 응답(서버 타임아웃 180초 → refusal no_knowledge/timeout). 같은 request_id 는 캐시 응답.
    `feedback[].reason` 은 사람 입력으로 신뢰되어 검열 규칙(learned.yaml)으로 되먹임된다. 인증: `Authorization: Bearer <RFA_ASK_TOKEN>`
    (.env.dev).'
  version: 1.0.0
paths:
...
  /inbox:
    get:
      tags:
      - inbox
      summary: Inbox List
      description: '결재함 목록: RFA_module 결재(:8790) + rfa_mas 접수 기록(작성 중 포함), 최신순. guest 는 사외 항목만(D-23).
        status: all | needs_approval | drafting | pending | regenerating | posting | publish_failed |
        posted | closed.'
      operationId: inboxList
      parameters:
      - name: task
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Task
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 32
          - type: 'null'
          title: Status
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InboxItem'
                title: Response Inboxlist
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Unprocessable Entity
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Service Unavailable
  /inbox/summary:
    get:
      tags:
      - inbox
      summary: Inbox Summary
      description: 결재 필요 수(결재함 배지·「결재 필요 N」), 태스크별 결재 필요 수, 상태별 수.
      operationId: inboxSummary
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InboxSummary'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Unprocessable Entity
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Service Unavailable
  /inbox/{item_id}:
    get:
      tags:
      - inbox
      summary: Inbox Detail
      description: '결재 상세: 원본 화면, 주입 의심 문장, 초안 카드(단계 RAG 검색·검증·LLM 초안, 재생성 이력, 게시 결과), 기록.'
      operationId: inboxDetail
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          title: Item Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InboxDetail'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Unprocessable Entity
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Service Unavailable
  /inbox/{item_id}/respond:
    post:
      tags:
      - inbox
      summary: Inbox Respond
      description: 바로 응답 = RFA_module approve(게시). owner. 화면의 초안이 저장된 초안과 다르면 409 `draft_edit_unsupported`(D-22).
        게시 실패는 502, 다시 부르면 재시도.
      operationId: inboxRespond
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          title: Item Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/RespondIn'
              - type: 'null'
              title: Body
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InboxDetail'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Unprocessable Entity
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Service Unavailable
  /inbox/{item_id}/regenerate:
    post:
      tags:
      - inbox
      summary: Inbox Regenerate
      description: 재생성 요청 = RFA_module reject(사유 = 요청 내용). owner. 대응 에이전트가 다시 써서 round+1 로 올린다. 3번째 요청이면
        결재가 닫힌다(응답하지 않기).
      operationId: inboxRegenerate
      parameters:
      - name: item_id
        in: path
        required: true
        schema:
          type: string
          title: Item Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegenerateIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InboxDetail'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Unprocessable Entity
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Service Unavailable
  /tasks/{task_id}/sources:
    get:
      tags:
      - inbox
      - inbox
      summary: List Sources
      description: 태스크 에이전트의 소스(문의가 들어오는 곳).
      operationId: listTaskSources
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SourceOut'
                title: Response Listtasksources
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Unprocessable Entity
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Service Unavailable
    post:
      tags:
      - inbox
      summary: Add Source
      description: '소스 연결(owner): GitHub `owner/repo` 또는 Slack `#채널`·`DM · 팀이름`, 받을 범위, 기본 등급. 이 소스로 들어온
        요청은 이 등급·태스크를 쓴다(D-24).'
      operationId: addTaskSource
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SourceIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SourceOut'
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Unprocessable Entity
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Service Unavailable
  /tasks/{task_id}/sources/{source_id}:
    delete:
      tags:
      - inbox
      summary: Remove Source
      description: 소스 떼기(owner).
      operationId: removeTaskSource
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      - name: source_id
        in: path
        required: true
        schema:
          type: string
          title: Source Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '204':
          description: Successful Response
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Forbidden
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Conflict
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Unprocessable Entity
        '502':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Bad Gateway
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorOut'
          description: Service Unavailable
...
components:
  schemas:
...
    Badge:
      properties:
        label:
          type: string
          title: Label
          description: 작성 중 | 초안 없음 | 결재 필요 | 차단됨 | 재생성 중 | 보내는 중 | 게시 실패 | 응답 완료 | 결재 완료
        tone:
          type: string
          enum:
          - info
          - warn
          - danger
          - ok
          - muted
          title: Tone
      type: object
      required:
      - label
      - tone
      title: Badge
    BlockedAttempt:
      properties:
        at:
          anyOf:
          - type: number
          - type: 'null'
          title: At
        kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
        action:
          type: string
          title: Action
        reason:
          type: string
          title: Reason
      type: object
      required:
      - at
      - kind
      - action
      - reason
      title: BlockedAttempt
...
    Color:
      properties:
        bg:
          type: string
          title: Bg
          description: '#rrggbb background'
        fg:
          type: string
          title: Fg
          description: '#rrggbb foreground'
      type: object
      required:
      - bg
      - fg
      title: Color
...
    DraftView:
      properties:
        text:
          type: string
          title: Text
        chars:
          type: integer
          title: Chars
        round:
          type: integer
          title: Round
        phase:
          type: string
          enum:
          - drafting
          - stalled
          - ready
          - regenerating
          - posting
          - publish_failed
          - posted
          - closed
          title: Phase
        regenerations:
          items:
            $ref: '#/components/schemas/Regeneration'
          type: array
          title: Regenerations
        lastRequest:
          anyOf:
          - type: string
          - type: 'null'
          title: Lastrequest
        postedUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Postedurl
        publishError:
          anyOf:
          - type: string
          - type: 'null'
          title: Publisherror
        canRespond:
          type: boolean
          title: Canrespond
        canRegenerate:
          type: boolean
          title: Canregenerate
        regenerationsLeft:
          type: integer
          title: Regenerationsleft
        closesOnRegenerate:
          type: boolean
          title: Closesonregenerate
          description: RFA_module closes the approval on its 3rd rejection
      type: object
      required:
      - text
      - chars
      - round
      - phase
      - regenerations
      - lastRequest
      - postedUrl
      - publishError
      - canRespond
      - canRegenerate
      - regenerationsLeft
      - closesOnRegenerate
      title: DraftView
...
    ErrorOut:
      properties:
        code:
          type: string
          title: Code
        message:
          type: string
          title: Message
      type: object
      required:
      - code
      - message
      title: ErrorOut
      description: 'Front-end error body: ``code`` for the client, ``message`` ready to show (Korean).'
...
    InboxAgent:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        desk:
          type: string
          title: Desk
          description: agent tag text (e.g. infer-opt)
        icon:
          type: string
          title: Icon
        color:
          $ref: '#/components/schemas/Color'
        initials:
          type: string
          title: Initials
      type: object
      required:
      - id
      - name
      - desk
      - icon
      - color
      - initials
      title: InboxAgent
    InboxDetail:
      properties:
        id:
          type: string
          title: Id
          description: stable per mention URL (before and after the approval exists)
        approvalId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Approvalid
          description: RFA_module approval id (결재 N); null while drafting
        sourceUrl:
          type: string
          title: Sourceurl
        channel:
          type: string
          title: Channel
        channelLabel:
          type: string
          title: Channellabel
        target:
          anyOf:
          - type: string
          - type: 'null'
          title: Target
        requester:
          type: string
          title: Requester
        requesterInitials:
          type: string
          title: Requesterinitials
        title:
          type: string
          title: Title
        grade:
          type: string
          enum:
          - public
          - company
          title: Grade
        gradeLabel:
          type: string
          title: Gradelabel
          description: 사외 | 사내
        gradeSource:
          type: string
          enum:
          - source
          - channel
          title: Gradesource
        status:
          type: string
          enum:
          - drafting
          - stalled
          - pending
          - regenerating
          - posting
          - publish_failed
          - posted
          - closed
          title: Status
        badge:
          $ref: '#/components/schemas/Badge'
        statusLine:
          type: string
          title: Statusline
        round:
          type: integer
          title: Round
        task:
          anyOf:
          - $ref: '#/components/schemas/InboxTask'
          - type: 'null'
        agent:
          anyOf:
          - $ref: '#/components/schemas/InboxAgent'
          - type: 'null'
        injection:
          anyOf:
          - $ref: '#/components/schemas/InjectionView'
          - type: 'null'
        arrivedAt:
          anyOf:
          - type: number
          - type: 'null'
          title: Arrivedat
        updatedAt:
          anyOf:
          - type: number
          - type: 'null'
          title: Updatedat
        question:
          type: string
          title: Question
        source:
          $ref: '#/components/schemas/SourceView'
        refusal:
          anyOf:
          - type: string
          - type: 'null'
          title: Refusal
        draft:
          $ref: '#/components/schemas/DraftView'
        steps:
          items:
            $ref: '#/components/schemas/Step'
          type: array
          title: Steps
        blockedAttempts:
          items:
            $ref: '#/components/schemas/BlockedAttempt'
          type: array
          title: Blockedattempts
        events:
          items:
            $ref: '#/components/schemas/ApprovalEvent'
          type: array
          title: Events
      type: object
      required:
      - id
      - approvalId
      - sourceUrl
      - channel
      - channelLabel
      - target
      - requester
      - requesterInitials
      - title
      - grade
      - gradeLabel
      - gradeSource
      - status
      - badge
      - statusLine
      - round
      - task
      - agent
      - injection
      - arrivedAt
      - updatedAt
      - question
      - source
      - refusal
      - draft
      - steps
      - blockedAttempts
      - events
      title: InboxDetail
    InboxItem:
      properties:
        id:
          type: string
          title: Id
          description: stable per mention URL (before and after the approval exists)
        approvalId:
          anyOf:
          - type: integer
          - type: 'null'
          title: Approvalid
          description: RFA_module approval id (결재 N); null while drafting
        sourceUrl:
          type: string
          title: Sourceurl
        channel:
          type: string
          title: Channel
        channelLabel:
          type: string
          title: Channellabel
        target:
          anyOf:
          - type: string
          - type: 'null'
          title: Target
        requester:
          type: string
          title: Requester
        requesterInitials:
          type: string
          title: Requesterinitials
        title:
          type: string
          title: Title
        grade:
          type: string
          enum:
          - public
          - company
          title: Grade
        gradeLabel:
          type: string
          title: Gradelabel
          description: 사외 | 사내
        gradeSource:
          type: string
          enum:
          - source
          - channel
          title: Gradesource
        status:
          type: string
          enum:
          - drafting
          - stalled
          - pending
          - regenerating
          - posting
          - publish_failed
          - posted
          - closed
          title: Status
        badge:
          $ref: '#/components/schemas/Badge'
        statusLine:
          type: string
          title: Statusline
        round:
          type: integer
          title: Round
        task:
          anyOf:
          - $ref: '#/components/schemas/InboxTask'
          - type: 'null'
        agent:
          anyOf:
          - $ref: '#/components/schemas/InboxAgent'
          - type: 'null'
        injection:
          type: boolean
          title: Injection
        arrivedAt:
          anyOf:
          - type: number
          - type: 'null'
          title: Arrivedat
        updatedAt:
          anyOf:
          - type: number
          - type: 'null'
          title: Updatedat
      type: object
      required:
      - id
      - approvalId
      - sourceUrl
      - channel
      - channelLabel
      - target
      - requester
      - requesterInitials
      - title
      - grade
      - gradeLabel
      - gradeSource
      - status
      - badge
      - statusLine
      - round
      - task
      - agent
      - injection
      - arrivedAt
      - updatedAt
      title: InboxItem
    InboxSummary:
      properties:
        needsApproval:
          type: integer
          title: Needsapproval
          description: 「결재 필요 N」, 결재함 badge
        byTask:
          additionalProperties:
            type: integer
          type: object
          title: Bytask
          description: pending per task id (left-nav badges)
        byStatus:
          additionalProperties:
            type: integer
          type: object
          title: Bystatus
        autoReplied:
          type: integer
          title: Autoreplied
          default: 0
        total:
          type: integer
          title: Total
      type: object
      required:
      - needsApproval
      - byTask
      - byStatus
      - total
      title: InboxSummary
    InboxTask:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
      - id
      - name
      title: InboxTask
...
    InjectionView:
      properties:
        sentences:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Sentences
        note:
          type: string
          title: Note
      type: object
      required:
      - sentences
      - note
      title: InjectionView
...
    RegenerateIn:
      properties:
        request:
          type: string
          maxLength: 400
          title: Request
          description: 재생성 요청 내용 (RFA_module reject 사유로 전달)
      type: object
      required:
      - request
      title: RegenerateIn
...
    RespondIn:
      properties:
        draft:
          anyOf:
          - type: string
          - type: 'null'
          title: Draft
          description: the text on screen; must equal the stored draft (D-22)
      type: object
      title: RespondIn
...
    SourceIn:
      properties:
        kind:
          type: string
          enum:
          - github
          - slack
          title: Kind
        target:
          type: string
          maxLength: 200
          title: Target
        scopes:
          items:
            type: string
          type: array
          maxItems: 6
          title: Scopes
        grade:
          anyOf:
          - type: string
          - type: 'null'
          title: Grade
          description: public|company (사외|사내); default github 사외, slack 사내
      type: object
      required:
      - kind
      - target
      title: SourceIn
    SourceOut:
      properties:
        id:
          type: string
          title: Id
        taskId:
          type: string
          title: Taskid
        kind:
          type: string
          enum:
          - github
          - slack
          title: Kind
        kindLabel:
          type: string
          title: Kindlabel
        target:
          type: string
          title: Target
        scopes:
          items:
            type: string
          type: array
          title: Scopes
        grade:
          type: string
          enum:
          - public
          - company
          title: Grade
        gradeLabel:
          type: string
          title: Gradelabel
        status:
          type: string
          const: connected
          title: Status
        createdAt:
          type: number
          title: Createdat
        requestCount:
          type: integer
          title: Requestcount
        lastRequestAt:
          anyOf:
          - type: number
          - type: 'null'
          title: Lastrequestat
      type: object
      required:
      - id
      - taskId
      - kind
      - kindLabel
      - target
      - scopes
      - grade
      - gradeLabel
      - status
      - createdAt
      - requestCount
      - lastRequestAt
      title: SourceOut
...
    SourceView:
      properties:
        kind:
          type: string
          enum:
          - github
          - slack
          title: Kind
        url:
          type: string
          title: Url
      additionalProperties: true
      type: object
      required:
      - kind
      - url
      title: SourceView
      description: '원본 화면: GitHub (repo, number, title, body, comments) or Slack (channel, messages).'
    Step:
      properties:
        key:
          type: string
          enum:
          - rag
          - verify
          - draft
          title: Key
        title:
          type: string
          title: Title
        state:
          type: string
          enum:
          - pending
          - running
          - done
          - error
          - skipped
          - unknown
          title: State
        ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ms
        summary:
          type: string
          title: Summary
        details:
          items:
            $ref: '#/components/schemas/StepDetail'
          type: array
          title: Details
      type: object
      required:
      - key
      - title
      - state
      - ms
      - summary
      - details
      title: Step
...